move FogbenderSimpleWidget into its own file
